Define frameshift regex as raw string to prevent Syntax Warnings
In more recent versions of Python 3, when the string "\*" is encountered, it raises a SyntaxWarning because it is not a valid escape sequence. To avoid that warning, we can define the string as raw by inserting an "r" before it. This way, Python does not attempt to interpret "\*" as an escape sequence, and the warning is no longer triggered. Important: The previous version of the code was working correctly because when Python cannot find "\*" in the escape sequence dictionary, it treats it as a raw string. This is the desired behavior but had the inconvenience of displaying the warning message.